fs: add support for custom fsx_xflags_mask
With getfsxattrat() syscall, filesystem may use this field to report
its supported xflags.  Zero mask value means that supported flags are
not advertized.

With setfsxattrat() syscall, userspace may use this field to declare
which xflags and fields are being set.  Zero mask value means that
all known xflags and fields are being set.

Programs that call getfsxattrat() to fill struct fsxattr before calling
setfsxattrat() will not be affected by this change, but it allows
programs that call setfsxattrat() without calling getfsxattrat() to make
changes to some xflags and fields without knowing or changing the values
of unrelated xflags and fields.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>