fanotify: generate pre-dir-content event once on readdir
Unlike the legacy FAN_ACCESS_PERM event, the FAN_PRE_ACCESS pre-content
event is not allowed with FAN_ONDIR, so does not generate events on
directories.

Introduce a new pre-dir-content event which is called before first access
to directory content to add the ability to populate a directory before
readdir.

When this event is handled, the DCACHE_HSM_ONCE flag is set on the dentry
to suppress further pre-dir-content hooks as readdir continues.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>