fsnotify: fail some negative lookup results for internal vfs users
Internal vfs lookup users, like overlayfs, cannot guarantee that the
pre-content lookup hook is called from a safe context without fs locks,
so we cannot block on userspace response for those calls.

Instead, we always allow the lookup and we only call the hook if the
lookup result is negative and if the lookup should have be handled in
usersapce, the hook return -ENOENT.

In that case, the negative dentry is dropped and error -ENOENT is
returned to the caller.

This allows internal vfs users (e.g. overlayfs, nfsd) to access and
operate on files in HSM populated directories, but does not allow them
to create files inside an HSM un-populated directory.

Link: https://lore.kernel.org/linux-fsdevel/CAOQ4uxhttUt6makW6GZGfBb=rat+gH9QQmparX3cexDJwzhVMw@mail.gmail.com/
Signed-off-by: Amir Goldstein <amir73il@gmail.com>