fanotify: generate pre-dir-content event with name info on lookup
The pre-dir-content event on readdir can be used to populate an entire
directory before an application is reading its content.

This new hook adds the ability to populate an entire directory or a
single directory entry before lookup of a child name in the directory.

The hook is called from lookup_slow_notify() when travesing a directory
during path lookup walk.

The pre-dir-content event cannot be sent to usersapce while lookup is in
RCU walk, so if dcache entry already exist, it is assumed that the
pre-dir-content event is not needed, because an event was already
handled when dcache was populated.

If the event was generated from readdir before lookup, the
DCACHE_HSM_ONCE flag will have been set and this lookup event will not
be generated, but the lookup event does not set this flag on success,
because this event might have been handled by populating a single entry.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>