fsnotify: prepare for pre-dir-content events
The group mode of FAN_CLASS_PRE_CONTENT_FID is designated for add
support for pre-content events on directories.

Define the event FS_PRE_DIR_ACCESS and split the macro
FANOTIFY_PRE_CONTENT_EVENTS of pre-content events into
FANOTIFY_PRE_{DIR,FILE}_CONTENT_EVENTS.

Pre-content events on directories will not be generated for subdirs
when watching the parent directory, so we do not need to check the
parent's mark mask.

Create the macro FANOTIFY_DIRONLY_EVENTS for events that are only
reported on directories and use it instead of the more specific
macro ALL_FSNOTIFY_DIRENT_EVENTS, which were the only events that
fell into this category until now.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>