fsnotify: optimize out multiple pre-dir-content events during path walk
During lookup, the common case is that there are no pre-dir-content
events on any of the filesystem objects or that the filesystem does
not support pre-content events at all.

Identify this case by returning -EOPNOTSUPP error from
fsnotify_lookup_perm(), so that path lookup can skip calling the hook
until it jumps to another mount, which is possibly another sb.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>