fs: factor out helper lookup_drop_negative()
The drop_negative argument is used in overlayfs to avoid leaving a
negative dentry in dcache as a result of a positive only lookup.

Hoist this functionality to the lookup_one_positive_unlocked helper,
so more users can use it and factor out an internal helper needed
by other lookup helpers that do not wish to leave a negative dentry
behind.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>